perf: parallelize functional tests and split MonorepoPluginFunctionalTest
- Add maxParallelForks to functionalTest task (CPUs/2) so Gradle distributes
  test classes across forked JVMs
- Increase heap to 4GB to support multiple concurrent test JVMs
- Add --parallel to all GradleRunner invocations so inner test builds can
  configure/execute subprojects in parallel
- Split MonorepoPluginFunctionalTest (21 tests) into three focused classes
  to give maxParallelForks more units of work to distribute:
    - MonorepoPluginDetectionFunctionalTest: basic committed/staged/untracked detection
    - MonorepoPluginNestedProjectFunctionalTest: three-level-deep project hierarchies
    - MonorepoPluginHierarchyNodeFunctionalTest: intermediate directory node exclusion
- Delete empty MonorepoPluginEdgeCasesTest
